test: DRY out testing
Also, pulls out the 'wait-port' invocation to the root. Reason being,
each workspace should be able to work without relying on anything
external to it. However, those test don't start the server and so cannot
be called in isolation.

Also, this is a touch drier and makes it easy to run the startup tests
in parallel.